refactor: cleaned the code and removed useless header calls

+++ src/main_menu/main_menu.c
@@ -10,7 +10,6 @@
 #include <time.h>
 #include "my_game.h"
 #include "my.h"
-#include "score.h"
 
 void draw_active_navbar_line(game_data_t *game,
     float pos_x, float size_x)
@@ -27,7 +26,7 @@ void draw_active_navbar_line(game_data_t *game,
 void basic_menu(game_data_t *game)
 {
     sprite_id_t elements[] = {SP_MAIN_BG, SP_VENDETTA_MENU};
-    int element_count = 2;
+    size_t element_count = 2;
 
     for (size_t i = 0; i < element_count; i++) {
         sfRenderWindow_drawSprite(game->window,
